+++ docs/implementors/core/convert/trait.From.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {};
-implementors["my_iot"] = [{text:"impl&lt;'_, '_&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;'_ Row&lt;'_&gt;&gt; for <a class=\"struct\" href=\"my_iot/reading/struct.Reading.html\" title=\"struct my_iot::reading::Reading\">Reading</a>",synthetic:false,types:["my_iot::reading::Reading"]},];
+implementors["my_iot"] = [{text:"impl&lt;'_, '_&gt; <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/convert/trait.From.html\" title=\"trait core::convert::From\">From</a>&lt;&amp;'_ Row&lt;'_&gt;&gt; for <a class=\"struct\" href=\"my_iot/message/struct.Reading.html\" title=\"struct my_iot::message::Reading\">Reading</a>",synthetic:false,types:["my_iot::message::Reading"]},];
 
             if (window.register_implementors) {
                 window.register_implementors(implementors);