+++ src/services/buienradar.rs
@@ -1,6 +1,6 @@
 use crate::consts::USER_AGENT;
 use crate::db::Db;
-use crate::reading::{Message, Reading, Type};
+use crate::message::{Message, Reading, Type};
 use crate::services::Service;
 use crate::threading;
 use crate::value::{PointOfTheCompass, Value};